docs: update select sheet ordering
